adding functions

+++ acquire.py
@@ -14,21 +14,43 @@ def get_df(name):
     api_url = base_url + '/api/v1/'
     response = requests.get(api_url + name)
     data = response.json()
+
+    file_name=(name+'.csv')
+
+    if os.path.isfile(file_name):
+        return pd.read_csv(name+'.csv')
+    else:
+        # create list from 1st page
+        my_list = data['payload'][name]
+        
+        # loop through the pages and add to list
+        while data['payload']['next_page'] != None:
+            response = requests.get(base_url + data['payload']['next_page'])
+            data = response.json()
+            my_list.extend(data['payload'][name])
+        
+        # Create DataFrame from list
+        df = pd.DataFrame(my_list)
+        
+        # Write DataFrame to csv file for future use
+        df.to_csv(name + '.csv')
+
+    return df
+
+
+def combine_df(items, sales, stores):
+    '''
+    This functions takes in the three dataframes, items, sales, and stores and merges them.
+    '''
     
-    # create list from 1st page
-    my_list = data['payload'][name]
+    # rename columns to have a primary key
+    items.rename(columns={'item_id':'item'}, inplace=True)
+    stores.rename(columns={'store_id':'store'}, inplace=True)
     
-    # loop through the pages and add to list
-    while data['payload']['next_page'] != None:
-        response = requests.get(base_url + data['payload']['next_page'])
-        data = response.json()
-        my_list.extend(data['payload'][name])
+    # merge the dataframes together
+    items_sales = items.merge(sales, how='right', on='item')
+    df = items_sales.merge(stores, how='left', on='store')
     
-    # Create DataFrame from list
-    df = pd.DataFrame(my_list)
-    
-    # Write DataFrame to csv file for future use
-    df.to_csv(name + '.csv')
     return df
